Add initial version of notes_for_git_branch.py.
This is a basic proof of concept for notes_for_git_branch.
The idea is that the script opens a notes file for the current branch of
the current git repository. This then making it easy to keep notes for
the current task branch.

The branch and respective toplevel can be provided via command line
options as well. The directory for the notes is expected as an
environment variable.