Make editor configurable.
For some reason not everyone uses vim. For this reason we now allow to
specify an editor as commandline parameter. Alternatively it tries to
use the one in '$EDITOR', or finally vi. A lot of systems link vi to
vim, but it should still be more common than vim on default
installations.