Add --list option to print all existing notes.
This options helps to determine which notes currently exist. Because
otherwise short of actually checking NOTES_DIR there is no way of
determining it.

For now it simply prints all notes and disregards all other provided
arguments. We may consider to allow specifying a specific toplevel so
that only the notes from that toplevel are printed.