Use os.path.join instead of manually joining paths.
Concatenating paths by manually inserting '/' works fine on unix, but
may not work on Windows where '\' is used as a path separator. Using
os.path.join is the correct way to approach this as it selects the
correct path separator of the respective platform and makes sure that
absolute paths are correctly joined together.