Ignore hidden files in --list view.
Whilst notes are open vim for example creates a hidden temporary file.
We do not want to show that one though. So we simply ignore hidden files
altogether.

We may consider checking against the file ending as well to make sure
that we actually only list notes and not other files as well.