Split editor command into a list of strings.
subprocess does the necessary escaping of command line arguments to
calls by itself, but it requires a list of strings to do so. For example
'vim --noplugin' needs to be provided as ['vim', '--noplugin'], as
otherwise it tries to run it as one command. We simply split the command
that we get to achieve this.

This allows to specify commandline options together with the command and
it just works as expected.