Use 'open' and 'list' commands to determine branch-notes behaviour.
This means that instead of calling 'branch-notes' it is necessary to
call 'branch-notes open' to open a notes file. Listing notes turns from
'branch-notes --list' to 'branch-notes list'.
The reason for this is that it makes it easier to implement the
bash autocompletion as we can build on the different action commands.