Refactor getting of subprocess output into its own function.
This does not really save a lot of lines of code (it actually adds lines
overall), but it simply avoids having to implement the special handling
of for example using the UTF-8 encoding and calling rstrip twice.
If we ever change that logic again we will only have to do so in one
place.