Search for notes file if branch is explicitly specified on commandline.
Instead of having to specify a toplevel we now simply search for the
notes file if a branch has been given on the commandline. Now for
example `branch-notes open master` will search for `master.txt` in the
notes directory and open the respective file. If more than one instance
exists it will print an error and the user has to provide the toplevel.

As before, when no branch is specified we simply use the one from the
current directory.

The only problem with these changes is that it is not possible anymore
to open a specific branch from the current toplevel. This seems neglible
right now though.