Add option to archive a note.
It turns out that creating lots of notes results in a lot of files. This
makes the list command less helpful and opening specific notes less
intuitive. One solution is to delete unused notes. This is easily done,
but it would be good to preserve them instead. For this we implement the
archive command. We simply move the desired from its current location to
the same toplevel directory under NOTES_DIR/ARCHIVE.

Note that archived notes somewhat implicitly are not found by list and
open commands, as they are two directory levels deep in the archive.