Add autocomplete support for archive command.
Basically it does what one would expect. It suggests existing notes to
be archived and allows to set the toplevel to address a specific one.

Note that allowed parameters for archive are the same as for the list
command. It seems more explicit and more clear to copy those instead of
using the same variable or block in the switch-case statement.