Rename BRANCH_NOTES_EDITOR env variable to NOTES_EDITOR.
This is a workaround for the issue that zsh with case-insensitive
completion completes 'branch' to 'BRANCH_NOTES_EDITOR, branch-notes'.
This is somewhat annoying as the default completion should be
branch-notes. After searching for a while and not finding a quick and
simple solution this seems like the easiest approach.

Also, the directory env variable is called NOTES_DIR so now they both
match.