Escape whitespaces in directory names for toplevel tab completion.
Unescaped they are seen as several completion options. For example
'directory a' would be completed as 'directory' and 'a' rather than just
'directory a'.

There is a fair bit of code duplication here with the logic to determine
the directories, which should ideally be refactored and merged together.
For now this works though.